Move all the Stripe logic into the premium module
